P3_DOC Document
This document consists of several requirements for Project 3. It has the User Guide embedded, with both text and screenshots to serve as a guide to the user to utilize the various functionalities of the application. It also has the location of the test cases located in our program. Furthermore, it has the name of the design pattern applied in parts of our programs and the location in which it was applied.
NOTE: This document was collaborated on by all the group members. It was uploaded by one of us solely because it is a document, and commits on it cant be shown in the history like code.
TO GROUP MEMBERS: we still need the location of the test cases as it is not present as of now. 😃